requirements.txt
APScheduler==3.6.3
attrs==22.1.0
bcrypt==4.0.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
distro==1.7.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jsonschema==3.2.0
packaging==21.3
paramiko==2.11.0
Pillow==9.2.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
tesseract==0.1.3
texttable==1.6.4
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
websocket-client==0.59.0
